Fix Undo a move of an nested state
Warning:
This will need Kevin's review :)

Background:
I just didn't got why the reanchoring/reparenting is needed. It
works fine here without that. So, trying this solution first
because in any case I would need need to know what this is for
to proper test any alternate patch.

What happens:
The reparenting actually results in the item being moved first
to the same x/y it had under its parent but under the new parent
means it jumps. This results in a x/y change which is recorded
and later, on undo, restored to.

I know that according to the ParentChange documentation this
shouldn't happen but it does.

This solves FOCA-88

Change-Id: I21e8dbe5e53e14fcd7243fa13ee3c6aeeee2e8e7
Reviewed-on: https://codereview.kdab.com/11823
Reviewed-by: Continuous Integration <build@kdab.com>
Reviewed-by: Kevin Funk <kevin.funk@kdab.com>