FOCA-137: Fix adjusting of state labels
Prior to this patch, a change of the state label could cause the visual
representation of this element to loose its initial geometry. This patch
now no longer recalculates the geometry of a state layout item when the
label got changed. Leave that to the layouter if needed.

Initially, that code was there to be able to get a somewhat useful
visual appearance of the state chart when no layouter was present.

Change-Id: Idf3d2bdb961b021220905d69285fdbddcd17b337