Fix Birds-eye view doesn't handle moving of the state machine
Problem:

If moving the state machine away from the top left corder (e.g.
to the center of the view) via drag and drop, the birdseye view
updates accordingly but doesn't resize itself. So you only see
left upper part of the view only partially covering the machine,
also prevent navigating over the full state machine using it.

This is not really a BirdEyeView only problem because even the
scrollbars do not show up correctly.

Reason:

We only took the root-LayoutItem's width/height but not its
position into account when defining the scene's width/height.

Change-Id: Idbb494028b793079395f7112616649a1103ab576
Reviewed-on: https://codereview.kdab.com/13366
Reviewed-by: Continuous Integration <build@kdab.com>
Reviewed-by: Volker Krause <volker@kdab.com>