cmake: Work-around duplicate symbol issues
Before this patch the linking step failed:
  ld: error: duplicate symbol: TrafficLightStateMachine::qt_metacall(QMetaObject::Call, int, void**)

This is due to the CMP0071 being set to NEW (CMake AUTOMOC starts to processed
GENERATED files as well).

Due to a leftover in qt5_add_statecharts() we need to manually specify
the files CMake AUTOMOC should exclude created files generated by that
function.